langchain-demo/requirements.txt
# 核心依赖
python-dotenv>=1.0.0

# LangChain 相关
langchain>=0.1.0
langchain-community>=0.0.20

# Langfuse 观测
langfuse>=2.0.0

# HTTP 请求
requests>=2.28.0

# 可选：其他有用的库
typing-extensions>=4.0.0 